feat: Implement DrawTask wrapper and DrawTaskType enum (Issue #18)
- Added draw/draw_task.h and draw/draw_task.cpp
- Defined scoped enum DrawTaskType mapping to LV_DRAW_TASK_TYPE_*
- Implemented DrawTask class wrapper
- Updated draw/draw.h and draw/draw.cpp to use new types
- Added tests/test_draw_task.cpp for verification
- Fixed compilation error in Object::event_proxy
- Updated CMakeLists.txt to disable demos/examples for faster build
