+++ tests/lv_conf_test.h
@@ -4,6 +4,9 @@
 #define LV_COLOR_DEPTH 32
 #define LV_MEMORY_SIZE (4096 * 1024)
 
+#define LV_USE_DEMO 0
+#define LV_BUILD_EXAMPLES 0
+
 #define LV_USE_BUTTON 1
 #define LV_USE_LABEL 1
 #define LV_USE_SLIDER 1
